matching CalculateManaRate, ItemMaxMana, Spellcraft, ItemSkill/LevelLimit activation requirement, and ItemDifficulty / arcane lore requirement up to eor data
