+++ Source/ACE.Server/Factories/LootGenerationFactory_Magic.cs
@@ -1,451 +1,799 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
 using ACE.Common;
 using ACE.Database.Models.World;
-using ACE.Database;
 using ACE.Entity.Enum;
+using ACE.Entity.Models;
 using ACE.Server.Factories.Entity;
 using ACE.Server.Factories.Tables;
+using ACE.Server.Managers;
 using ACE.Server.WorldObjects;
 
 namespace ACE.Server.Factories
 {
     public static partial class LootGenerationFactory
     {
-        /// <summary>
-        /// Creates Caster (Wand, Staff, Orb)
-        /// </summary>
-        public static WorldObject CreateCaster(TreasureDeath profile, bool isMagical, int wield = -1, bool forceWar = false, bool mutate = true)
+        private static void AssignMagic(WorldObject wo, TreasureDeath profile, TreasureRoll roll, bool isArmor = false)
         {
-            // Refactored 11/20/19  - HarliQ
-            int casterWeenie = 0;
-            int subType = 0;
-            int element = 0;
+            int numSpells = 0;
 
-            if (wield == -1)
-                wield = RollWieldDifficulty(profile.Tier, WieldType.Caster);
-
-            // Getting the caster Weenie needed.
-            if (wield == 0)
-            {
-                // Determine plain caster type: 0 - Orb, 1 - Sceptre, 2 - Staff, 3 - Wand
-                subType = ThreadSafeRandom.Next(0, 3);
-                casterWeenie = LootTables.CasterWeaponsMatrix[wield][subType];
-            }
-            else
-            {
-                // Determine caster type: 1 - Sceptre, 2 - Baton, 3 - Staff
-                int casterType = ThreadSafeRandom.Next(1, 3);
-
-                // Determine element type: 0 - Slashing, 1 - Piercing, 2 - Blunt, 3 - Frost, 4 - Fire, 5 - Acid, 6 - Electric, 7 - Nether
-                element = forceWar ? ThreadSafeRandom.Next(0, 6) : ThreadSafeRandom.Next(0, 7);
-                casterWeenie = LootTables.CasterWeaponsMatrix[casterType][element];
-            }
-
-            WorldObject wo = WorldObjectFactory.CreateNewWorldObject((uint)casterWeenie);
-
-            if (wo != null && mutate)
-                MutateCaster(wo, profile, isMagical, wield);
-
-            return wo;
-        }
-
-        private static void MutateCaster(WorldObject wo, TreasureDeath profile, bool isMagical, int? wieldDifficulty = null, TreasureRoll roll = null)
-        {
-            if (wieldDifficulty != null)
+            if (roll == null)
             {
                 // previous method
+                if (!AssignMagic_Spells(wo, profile, isArmor, out numSpells))
+                    return;
+            }
+            else
+            {
+                // new method
+                if (!AssignMagic_New(wo, profile, roll, out numSpells))
+                    return;
+            }
 
-                var wieldRequirement = WieldRequirement.RawSkill;
-                var wieldSkillType = Skill.None;
+            wo.UiEffects = UiEffects.Magical;
 
-                double elementalDamageMod = 0;
+            var maxBaseMana = GetMaxBaseMana(wo);
 
-                if (wieldDifficulty == 0)
+            wo.ManaRate = CalculateManaRate(maxBaseMana);
+
+            if (roll == null)
+            {
+                wo.ItemMaxMana = RollItemMaxMana(numSpells, profile.Tier);
+                wo.ItemCurMana = wo.ItemMaxMana;
+
+                wo.ItemSpellcraft = RollSpellcraft(wo);
+                wo.ItemDifficulty = RollItemDifficulty(wo);
+            }
+            else
+            {
+                wo.ItemMaxMana = RollItemMaxMana_New(wo, profile, roll, maxBaseMana);
+                wo.ItemCurMana = wo.ItemMaxMana;
+
+                wo.ItemSpellcraft = RollSpellcraft(wo, roll);
+
+                AddActivationRequirements(wo, roll);
+            }
+        }
+
+        private static bool AssignMagic_Spells(WorldObject wo, TreasureDeath profile, bool isArmor, out int numSpells)
+        {
+            SpellId[][] spells;
+            SpellId[][] cantrips;
+
+            int lowSpellTier = GetLowSpellTier(profile.Tier);
+            int highSpellTier = GetHighSpellTier(profile.Tier);
+
+            switch (wo.WeenieType)
+            {
+                case WeenieType.Clothing:
+                    spells = ArmorSpells.Table;
+                    cantrips = ArmorCantrips.Table;
+                    break;
+                case WeenieType.Caster:
+                    spells = WandSpells.Table;
+                    cantrips = WandCantrips.Table;
+                    break;
+                case WeenieType.Generic:
+                    spells = JewelrySpells.Table;
+                    cantrips = JewelryCantrips.Table;
+                    break;
+                case WeenieType.MeleeWeapon:
+                    spells = MeleeSpells.Table;
+                    cantrips = MeleeCantrips.Table;
+                    break;
+                case WeenieType.MissileLauncher:
+                    spells = MissileSpells.Table;
+                    cantrips = MissileCantrips.Table;
+                    break;
+                default:
+                    spells = null;
+                    cantrips = null;
+                    break;
+            }
+
+            if (wo.IsShield)
+            {
+                spells = ArmorSpells.Table;
+                cantrips = ArmorCantrips.Table;
+            }
+
+            numSpells = 0;
+
+            if (spells == null || cantrips == null)
+                return false;
+
+            // Refactor 3/2/2020 - HQ
+            // Magic stats
+            numSpells = GetSpellDistribution(profile, out int minorCantrips, out int majorCantrips, out int epicCantrips, out int legendaryCantrips);
+            int numCantrips = minorCantrips + majorCantrips + epicCantrips + legendaryCantrips;
+
+            int[] shuffledValues = Enumerable.Range(0, spells.Length).ToArray();
+
+            Shuffle(shuffledValues);
+
+            if (numSpells - numCantrips > 0)
+            {
+                for (int i = 0; i < numSpells - numCantrips; i++)
                 {
-                    if (profile.Tier > 6)
-                    {
-                        wieldRequirement = WieldRequirement.Level;
-                        wieldSkillType = Skill.Axe;  // Set by examples from PCAP data
+                    int col = ThreadSafeRandom.Next(lowSpellTier - 1, highSpellTier - 1);
+                    SpellId spellID = spells[shuffledValues[i]][col];
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
+                }
+            }
 
-                        wieldDifficulty = profile.Tier switch
-                        {
-                            7 => 150, // In this instance, used for indicating player level, rather than skill level
-                            _ => 180, // In this instance, used for indicating player level, rather than skill level
-                        };
+            // Per discord discussions: ALL armor/shields if it had any spells, had an Impen spell
+            if (isArmor)
+            {
+                var impenSpells = SpellLevelProgression.Impenetrability;
+
+                // Ensure that one of the Impen spells was not already added
+                bool impenFound = false;
+                for (int i = 0; i < 8; i++)
+                {
+                    if (wo.Biota.SpellIsKnown((int)impenSpells[i], wo.BiotaDatabaseLock))
+                    {
+                        impenFound = true;
+                        break;
                     }
                 }
-                else
+                if (!impenFound)
                 {
-                    elementalDamageMod = RollElementalDamageMod(wieldDifficulty.Value);
-
-                    if (wo.W_DamageType == DamageType.Nether)
-                        wieldSkillType = Skill.VoidMagic;
-                    else
-                        wieldSkillType = Skill.WarMagic;
+                    int col = ThreadSafeRandom.Next(lowSpellTier - 1, highSpellTier - 1);
+                    SpellId spellID = impenSpells[col];
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
                 }
-
-                // ManaConversionMod
-                var manaConversionMod = RollManaConversionMod(profile.Tier);
-                if (manaConversionMod > 0.0f)
-                    wo.ManaConversionMod = manaConversionMod;
-
-                // ElementalDamageMod
-                if (elementalDamageMod > 1.0f)
-                    wo.ElementalDamageMod = elementalDamageMod;
-
-                // WieldRequirements
-                if (wieldDifficulty > 0 || wieldRequirement == WieldRequirement.Level)
-                {
-                    wo.WieldRequirements = wieldRequirement;
-                    wo.WieldSkillType = (int)wieldSkillType;
-                    wo.WieldDifficulty = wieldDifficulty;
-                }
-                else
-                {
-                    wo.WieldRequirements = WieldRequirement.Invalid;
-                    wo.WieldSkillType = null;
-                    wo.WieldDifficulty = null;
-                }
-
-                // WeaponDefense
-                wo.WeaponDefense = RollWeaponDefense(wieldDifficulty.Value, profile);
             }
-            else
+
+            if (numCantrips > 0)
             {
-                // new method - mutation scripts
+                shuffledValues = Enumerable.Range(0, cantrips.Length).ToArray();
+                Shuffle(shuffledValues);
 
-                // mutate ManaConversionMod
-                var mutationFilter = MutationCache.GetMutation("Casters.caster.txt");
-                mutationFilter.TryMutate(wo, profile.Tier);
+                int shuffledPlace = 0;
 
-                // mutate ElementalDamageMod / WieldRequirements
-                var isElemental = wo.W_DamageType != DamageType.Undef;
-                var scriptName = GetCasterScript(isElemental);
-
-                mutationFilter = MutationCache.GetMutation(scriptName);
-                mutationFilter.TryMutate(wo, profile.Tier);
-
-                // this part was not handled by mutation filter
-                if (wo.WieldRequirements == WieldRequirement.RawSkill)
+                // minor cantrips
+                for (var i = 0; i < minorCantrips; i++)
                 {
-                    if (wo.W_DamageType == DamageType.Nether)
-                        wo.WieldSkillType = (int)Skill.VoidMagic;
-                    else
-                        wo.WieldSkillType = (int)Skill.WarMagic;
+                    SpellId spellID = cantrips[shuffledValues[shuffledPlace]][0];
+                    shuffledPlace++;
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
                 }
-
-                // mutate WeaponDefense
-                mutationFilter = MutationCache.GetMutation("Casters.weapon_defense.txt");
-                mutationFilter.TryMutate(wo, profile.Tier);
+                // major cantrips
+                for (var i = 0; i < majorCantrips; i++)
+                {
+                    SpellId spellID = cantrips[shuffledValues[shuffledPlace]][1];
+                    shuffledPlace++;
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
+                }
+                // epic cantrips
+                for (var i = 0; i < epicCantrips; i++)
+                {
+                    SpellId spellID = cantrips[shuffledValues[shuffledPlace]][2];
+                    shuffledPlace++;
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
+                }
+                // legendary cantrips
+                for (var i = 0; i < legendaryCantrips; i++)
+                {
+                    SpellId spellID = cantrips[shuffledValues[shuffledPlace]][3];
+                    shuffledPlace++;
+                    wo.Biota.GetOrAddKnownSpell((int)spellID, wo.BiotaDatabaseLock, out _);
+                }
             }
-
-            // material type
-            int materialType = GetMaterialType(wo, profile.Tier);
-            if (materialType > 0)
-                wo.MaterialType = (MaterialType)materialType;
-
-            // item color
-            MutateColor(wo);
-
-            // gem count / gem material
-            if (wo.GemCode != null)
-                wo.GemCount = GemCountChance.Roll(wo.GemCode.Value, profile.Tier);
-            else
-                wo.GemCount = ThreadSafeRandom.Next(1, 5);
-
-            wo.GemType = RollGemType(profile.Tier);
-
-            // workmanship
-            wo.ItemWorkmanship = GetWorkmanship(profile.Tier);
-
-            // burden?
-
-            // item value
-            wo.Value = GetValue(profile.Tier, wo.ItemWorkmanship.Value, LootTables.getMaterialValueModifier(wo), LootTables.getGemMaterialValueModifier(wo));
-
-            // missile defense / magic defense
-            wo.WeaponMissileDefense = RollWeapon_MissileMagicDefense(profile.Tier);
-            wo.WeaponMagicDefense = RollWeapon_MissileMagicDefense(profile.Tier);
-
-            // spells
-            if (!isMagical)
-            {
-                wo.ItemManaCost = null;
-                wo.ItemMaxMana = null;
-                wo.ItemCurMana = null;
-                wo.ItemSpellcraft = null;
-                wo.ItemDifficulty = null;
-            }
-            else
-            {
-                // if a caster was from a MagicItem profile, it always had a SpellDID
-                MutateCaster_SpellDID(wo, profile);
-
-                AssignMagic(wo, profile, roll);
-            }
-
-            // long description
-            wo.LongDesc = GetLongDesc(wo);
+            return true;
         }
 
-        private static void MutateCaster_SpellDID(WorldObject wo, TreasureDeath profile)
+        private static int GetSpellDistribution(TreasureDeath profile, out int numMinors, out int numMajors, out int numEpics, out int numLegendaries)
         {
-            var firstSpell = CasterSlotSpells.Roll(wo);
+            int numNonCantrips = 0;
 
-            var spellLevels = SpellLevelProgression.GetSpellLevels(firstSpell);
+            numMinors = 0;
+            numMajors = 0;
+            numEpics = 0;
+            numLegendaries = 0;
 
-            if (spellLevels == null)
-            {
-                log.Error($"MutateCaster_SpellDID: couldn't find {firstSpell}");
-                return;
-            }
+            int nonCantripChance = ThreadSafeRandom.Next(1, 100000);
 
-            if (spellLevels.Count != 8)
-            {
-                log.Error($"MutateCaster_SpellDID: found {spellLevels.Count} spell levels for {firstSpell}, expected 8");
-                return;
-            }
+            numMinors = GetNumMinorCantrips(profile); // All tiers have a chance for at least one minor cantrip
+            numMajors = GetNumMajorCantrips(profile);
+            numEpics = GetNumEpicCantrips(profile);
+            numLegendaries = GetNumLegendaryCantrips(profile);
 
-            int minSpellLevel = GetLowSpellTier(profile.Tier);
-            int maxSpellLevel = GetHighSpellTier(profile.Tier);
+            //  Fixing the absurd amount of spells on items - HQ 6/21/2020
+            //  From Mags Data all tiers have about the same chance for a given number of spells on items.  This is the ratio for magical items.
+            //  1 Spell(s) - 46.410 %
+            //  2 Spell(s) - 27.040 %
+            //  3 Spell(s) - 17.850 %
+            //  4 Spell(s) - 6.875 %
+            //  5 Spell(s) - 1.525 %
+            //  6 Spell(s) - 0.235 %
+            //  7 Spell(s) - 0.065 %
 
-            var spellLevel = ThreadSafeRandom.Next(minSpellLevel, maxSpellLevel);
+            if (nonCantripChance <= 46410)
+                numNonCantrips = 1;
+            else if (nonCantripChance <= 73450)
+                numNonCantrips = 2;
+            else if (nonCantripChance <= 91300)
+                numNonCantrips = 3;
+            else if (nonCantripChance <= 98175)
+                numNonCantrips = 4;
+            else if (nonCantripChance <= 99700)
+                numNonCantrips = 5;
+            else if (nonCantripChance <= 99935)
+                numNonCantrips = 6;
+            else
+                numNonCantrips = 7;
 
-            wo.SpellDID = (uint)spellLevels[spellLevel - 1];
-
-            var spell = new Server.Entity.Spell(wo.SpellDID.Value);
-
-            wo.ItemManaCost = (int)spell.BaseMana * 5;
-
-            wo.ItemUseable = Usable.SourceWieldedTargetRemoteNeverWalk;
+            return numNonCantrips + numMinors + numMajors + numEpics + numLegendaries;
         }
 
-        private static string GetCasterScript(bool isElemental = false)
+        private static int GetNumMinorCantrips(TreasureDeath profile)
         {
-            var elementalStr = isElemental ? "elemental" : "non_elemental";
+            int numMinors = 0;
 
-            return $"Casters.caster_{elementalStr}.txt";
+            var dropRate = PropertyManager.GetDouble("minor_cantrip_drop_rate").Item;
+            if (dropRate <= 0)
+                return 0;
+
+            var dropRateMod = 1.0 / dropRate;
+
+            double lootQualityMod = 1.0f;
+            if (PropertyManager.GetBool("loot_quality_mod").Item && profile.LootQualityMod > 0 && profile.LootQualityMod < 1)
+                lootQualityMod = 1.0f - profile.LootQualityMod;
+
+            switch (profile.Tier)
+            {
+                case 1:
+                    if (ThreadSafeRandom.Next(1, (int)(100 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 1;
+                    break;
+                case 2:
+                case 3:
+                    if (ThreadSafeRandom.Next(1, (int)(50 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(250 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 2;
+                    break;
+                case 4:
+                case 5:
+                    if (ThreadSafeRandom.Next(1, (int)(50 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(250 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 2;
+                    if (ThreadSafeRandom.Next(1, (int)(1000 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 3;
+                    break;
+                case 6:
+                case 7:
+                default:
+                    if (ThreadSafeRandom.Next(1, (int)(50 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(250 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 2;
+                    if (ThreadSafeRandom.Next(1, (int)(1000 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 3;
+                    if (ThreadSafeRandom.Next(1, (int)(5000 * dropRateMod * lootQualityMod)) == 1)
+                        numMinors = 4;
+                    break;
+            }
+
+            return numMinors;
         }
 
-        private static bool GetMutateCasterData(uint wcid)
+        private static int GetNumMajorCantrips(TreasureDeath profile)
         {
-            for (var i = 0; i < LootTables.CasterWeaponsMatrix.Length; i++)
-            {
-                var table = LootTables.CasterWeaponsMatrix[i];
+            int numMajors = 0;
 
-                for (var element = 0; element < table.Length; element++)
-                {
-                    if (wcid == table[element])
-                        return true;
-                }
+            var dropRate = PropertyManager.GetDouble("major_cantrip_drop_rate").Item;
+            if (dropRate <= 0)
+                return 0;
+
+            var dropRateMod = 1.0 / dropRate;
+
+            double lootQualityMod = 1.0f;
+            if (PropertyManager.GetBool("loot_quality_mod").Item && profile.LootQualityMod > 0 && profile.LootQualityMod < 1)
+                lootQualityMod = 1.0f - profile.LootQualityMod;
+
+            switch (profile.Tier)
+            {
+                case 1:
+                    numMajors = 0;
+                    break;
+                case 2:
+                    if (ThreadSafeRandom.Next(1, (int)(500 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 1;
+                    break;
+                case 3:
+                    if (ThreadSafeRandom.Next(1, (int)(500 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(10000 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 2;
+                    break;
+                case 4:
+                case 5:
+                case 6:
+                    if (ThreadSafeRandom.Next(1, (int)(500 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(5000 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 2;
+                    break;
+                case 7:
+                default:
+                    if (ThreadSafeRandom.Next(1, (int)(500 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 1;
+                    if (ThreadSafeRandom.Next(1, (int)(5000 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 2;
+                    if (ThreadSafeRandom.Next(1, (int)(15000 * dropRateMod * lootQualityMod)) == 1)
+                        numMajors = 3;
+                    break;
             }
-            return false;
+
+            return numMajors;
+        }
+
+        private static int GetNumEpicCantrips(TreasureDeath profile)
+        {
+            int numEpics = 0;
+
+            if (profile.Tier < 7)
+                return 0;
+
+            var dropRate = PropertyManager.GetDouble("epic_cantrip_drop_rate").Item;
+            if (dropRate <= 0)
+                return 0;
+
+            var dropRateMod = 1.0 / dropRate;
+
+            double lootQualityMod = 1.0f;
+            if (PropertyManager.GetBool("loot_quality_mod").Item && profile.LootQualityMod > 0 && profile.LootQualityMod < 1)
+                lootQualityMod = 1.0f - profile.LootQualityMod;
+
+            // 25% base chance for no epics for tier 7
+            if (ThreadSafeRandom.Next(1, 4) > 1)
+            {
+                // 1% chance for 1 Epic, 0.1% chance for 2 Epics,
+                // 0.01% chance for 3 Epics, 0.001% chance for 4 Epics 
+                if (ThreadSafeRandom.Next(1, (int)(100 * dropRateMod * lootQualityMod)) == 1)
+                    numEpics = 1;
+                if (ThreadSafeRandom.Next(1, (int)(1000 * dropRateMod * lootQualityMod)) == 1)
+                    numEpics = 2;
+                if (ThreadSafeRandom.Next(1, (int)(10000 * dropRateMod * lootQualityMod)) == 1)
+                    numEpics = 3;
+                if (ThreadSafeRandom.Next(1, (int)(100000 * dropRateMod * lootQualityMod)) == 1)
+                    numEpics = 4;
+            }
+
+            return numEpics;
+        }
+
+        private static int GetNumLegendaryCantrips(TreasureDeath profile)
+        {
+            int numLegendaries = 0;
+
+            if (profile.Tier < 8)
+                return 0;
+
+            var dropRate = PropertyManager.GetDouble("legendary_cantrip_drop_rate").Item;
+            if (dropRate <= 0)
+                return 0;
+
+            var dropRateMod = 1.0 / dropRate;
+
+            double lootQualityMod = 1.0f;
+            if (PropertyManager.GetBool("loot_quality_mod").Item && profile.LootQualityMod > 0 && profile.LootQualityMod < 1)
+                lootQualityMod = 1.0f - profile.LootQualityMod;
+
+            // 1% chance for a legendary, 0.02% chance for 2 legendaries
+            if (ThreadSafeRandom.Next(1, (int)(100 * dropRateMod * lootQualityMod)) == 1)
+                numLegendaries = 1;
+            if (ThreadSafeRandom.Next(1, (int)(500 * dropRateMod * lootQualityMod)) == 1)
+                numLegendaries = 2;
+
+            return numLegendaries;
+        }
+
+        private static int GetLowSpellTier(int tier)
+        {
+            int lowSpellTier;
+            switch (tier)
+            {
+                case 1:
+                    lowSpellTier = 1;
+                    break;
+                case 2:
+                    lowSpellTier = 3;
+                    break;
+                case 3:
+                    lowSpellTier = 4;
+                    break;
+                case 4:
+                    lowSpellTier = 5;
+                    break;
+                case 5:
+                case 6:
+                    lowSpellTier = 6;
+                    break;
+                default:
+                    lowSpellTier = 7;
+                    break;
+            }
+
+            return lowSpellTier;
+        }
+
+        private static int GetHighSpellTier(int tier)
+        {
+            int highSpellTier;
+            switch (tier)
+            {
+                case 1:
+                    highSpellTier = 3;
+                    break;
+                case 2:
+                    highSpellTier = 5;
+                    break;
+                case 3:
+                case 4:
+                    highSpellTier = 6;
+                    break;
+                case 5:
+                case 6:
+                    highSpellTier = 7;
+                    break;
+                default:
+                    highSpellTier = 8;
+                    break;
+            }
+
+            return highSpellTier;
+        }
+
+        private static void Shuffle<T>(T[] array)
+        {
+            // verified even distribution
+            for (var i = 0; i < array.Length; i++)
+            {
+                var idx = ThreadSafeRandom.Next(i, array.Length - 1);
+
+                var temp = array[idx];
+                array[idx] = array[i];
+                array[i] = temp;
+            }
         }
 
         /// <summary>
-        /// Rolls for ElementalDamageMod for caster weapons
+        /// Returns the maximum BaseMana from the spells in item's spellbook
         /// </summary>
-        private static double RollElementalDamageMod(int wield)
+        public static int GetMaxBaseMana(WorldObject wo)
         {
-            double elementBonus = 0;
+            var maxBaseMana = 0;
 
-            int chance = ThreadSafeRandom.Next(1, 100);
-            switch (wield)
+            if (wo.SpellDID != null)
             {
-                case 290:
-                    if (chance > 95)
-                        elementBonus = 0.03;
-                    else if (chance > 65)
-                        elementBonus = 0.02;
-                    else
-                        elementBonus = 0.01;
-                    break;
-                case 310:
-                    if (chance > 95)
-                        elementBonus = 0.06;
-                    else if (chance > 65)
-                        elementBonus = 0.05;
-                    else
-                        elementBonus = 0.04;
-                    break;
+                // added for ItemMaxMana calcs to match eor data
+                var spell = new Server.Entity.Spell(wo.SpellDID.Value);
 
-                case 330:
-                    if (chance > 95)
-                        elementBonus = 0.09;
-                    else if (chance > 65)
-                        elementBonus = 0.08;
-                    else
-                        elementBonus = 0.07;
-                    break;
+                if (spell.BaseMana > maxBaseMana)
+                    maxBaseMana = (int)spell.BaseMana;
+            }
 
-                case 355:
-                    if (chance > 95)
-                        elementBonus = 0.13;
-                    else if (chance > 80)
-                        elementBonus = 0.12;
-                    else if (chance > 55)
-                        elementBonus = 0.11;
-                    else if (chance > 20)
-                        elementBonus = 0.10;
-                    else
-                        elementBonus = 0.09;
-                    break;
+            if (wo.Biota.PropertiesSpellBook != null)
+            {
+                foreach (var spellId in wo.Biota.PropertiesSpellBook.Keys)
+                {
+                    var spell = new Server.Entity.Spell(spellId);
 
-                case 375:
-                    if (chance > 95)
-                        elementBonus = 0.16;
-                    else if (chance > 85)
-                        elementBonus = 0.15;
-                    else if (chance > 60)
-                        elementBonus = 0.14;
-                    else if (chance > 30)
-                        elementBonus = 0.13;
-                    else if (chance > 10)
-                        elementBonus = 0.12;
-                    else
-                        elementBonus = 0.11;
-                    break;
+                    if (spell.BaseMana > maxBaseMana)
+                        maxBaseMana = (int)spell.BaseMana;
+                }
+            }
+            return maxBaseMana;
+        }
 
-                default:
-                    // 385
-                    if (chance > 95)
-                        elementBonus = 0.18;
-                    else if (chance > 65)
-                        elementBonus = 0.17;
-                    else
-                        elementBonus = 0.16;
+        // old table / method
+
+        private static readonly List<(int min, int max)> itemMaxMana_RandomRange = new List<(int min, int max)>()
+        {
+            (200,  400),    // T1
+            (400,  600),    // T2
+            (600,  800),    // T3
+            (800,  1000),   // T4
+            (1000, 1200),   // T5
+            (1200, 1400),   // T6
+            (1400, 1600),   // T7
+            (1600, 1800),   // T8
+        };
+
+        private static int RollItemMaxMana(int tier, int numSpells)
+        {
+            var range = itemMaxMana_RandomRange[tier - 1];
+
+            var rng = ThreadSafeRandom.Next(range.min, range.max);
+
+            return rng * numSpells;
+        }
+
+        /// <summary>
+        /// Rolls the ItemMaxMana for an object
+        /// </summary>
+        private static int RollItemMaxMana_New(WorldObject wo, TreasureDeath profile, TreasureRoll roll, int maxBaseMana)
+        {
+            // verified matches up with magloot eor logs
+
+            var workmanship = WorkmanshipChance.GetModifier(wo.ItemWorkmanship - 1);
+
+            (int min, int max) range;
+
+            if (roll.IsClothing || roll.IsArmor || roll.IsMeleeWeapon || roll.IsMissileWeapon || roll.IsDinnerware)
+            {
+                range.min = 6;
+                range.max = 15;
+            }
+            else if (roll.IsCaster)
+            {
+                // casters seem to be the only oddity here
+                // they scale evenly per tier / with maxBaseMana still,
+                // however in a much larger 6-75 range, in a non-even distribution,
+                // unlike all of the other tables
+
+                // also verified these semi-rare casters were not outliers,
+                // ie. they all appeared to be standard lootgen casters
+
+                // this is most likely not exactly matched up to whatever retail was doing for casters,
+                // but at least approaches matching the possible range / distribution..
+
+                // the theoretical max (80 base mana * 2.0 ws mod * 75) would be 12k mana
+
+                range.min = 6;
+                range.max = 75;
+
+                var casterRange = range.max - range.min;
+
+                // roll a weighted value between 0-1
+                var interval = CasterItemMaxMana.Roll(profile);
+
+                var caster_rng = range.min + casterRange * interval;
+
+                return (int)Math.Ceiling(maxBaseMana * workmanship * caster_rng);
+
+            }
+            else if (roll.IsJewelry)
+            {
+                // includes crowns
+                range.min = 12;
+                range.max = 20;
+            }
+            else if (roll.IsGem)
+            {
+                range.min = 5;
+                range.max = 5;
+            }
+            else
+            {
+                log.Error($"RollItemMaxMana({wo.Name}, {roll.ItemType}, {maxBaseMana}) - unknown item type");
+                return 1;
+            }
+
+            var rng = ThreadSafeRandom.Next(range.min, range.max);
+
+            return (int)Math.Ceiling(maxBaseMana * workmanship * rng);
+        }
+
+        /// <summary>
+        /// Calculates the ManaRate for an item
+        /// </summary>
+        public static float CalculateManaRate(int maxBaseMana)
+        {
+            if (maxBaseMana <= 0)
+                maxBaseMana = 1;
+
+            // verified with eor data
+            return -1.0f / (float)Math.Ceiling(1200.0f / maxBaseMana);
+        }
+
+        // old method / based on item type
+
+        private static int RollSpellcraft(WorldObject wo)
+        {
+            var maxSpellPower = GetMaxSpellPower(wo);
+
+            (float min, float max) range = (1.0f, 1.0f);
+
+            switch (wo.ItemType)
+            {
+                case ItemType.Armor:
+                case ItemType.Clothing:
+                case ItemType.Jewelry:
+
+                case ItemType.MeleeWeapon:
+                case ItemType.MissileWeapon:
+                case ItemType.Caster:
+
+                    range = (0.9f, 1.1f);
                     break;
             }
 
-            elementBonus += 1;
+            var rng = ThreadSafeRandom.Next(range.min, range.max);
 
-            return elementBonus;
+            var spellcraft = (int)Math.Ceiling(maxSpellPower * rng);
+
+            // retail was capped at 370
+            spellcraft = Math.Min(spellcraft, 370);
+
+            return spellcraft;
         }
 
-        private static WorldObject CreateSummoningEssence(int tier, bool mutate = true)
+        // new method / based on treasure roll
+
+        private static int RollSpellcraft(WorldObject wo, TreasureRoll roll)
         {
-            uint id = 0;
+            var maxSpellPower = GetMaxSpellPower(wo);
 
-            // Adding a spread of Pet Device levels for each tier - Level 200 pets should only be dropping in T8 Loot - HQ 2/29/2020
-            // The spread is from Optim's Data
-            // T5-T8 20/35/30/15% split 
-            // T8- 200,180,150,125
-            // T7- 180,150,125,100
-            // T6- 150,125,100,80
-            // T5- 125,100,80,50
-            // T4- 100,80,50
-            // T3- 80,50
-            // T2- 50
-            // T1- 50
+            (float min, float max) range = (1.0f, 1.0f);
 
-            // Tables are already 1-7, so removing them being Tier dependent
-
-            int petLevel = 0;
-            int chance = ThreadSafeRandom.Next(1, 100);
-            if (chance > 80)
-                petLevel = tier - 1;
-            else if (chance > 45)
-                petLevel = tier - 2;
-            else if (chance > 15)
-                petLevel = tier - 3;
-            else
-                petLevel = tier - 4;
-            if (petLevel < 2)
-                petLevel = 1;
-
-            int summoningEssenceIndex = ThreadSafeRandom.Next(0, LootTables.SummoningEssencesMatrix.Length - 1);
-
-            id = (uint)LootTables.SummoningEssencesMatrix[summoningEssenceIndex][petLevel - 1];
-
-            var petDevice = WorldObjectFactory.CreateNewWorldObject(id) as PetDevice;
-
-            if (petDevice != null && mutate)
-                MutatePetDevice(petDevice, tier);
-
-            return petDevice;
-        }
-
-        private static void MutatePetDevice(WorldObject petDevice, int tier)
-        {
-            if (!(petDevice is PetDevice)) return;
-
-            var ratingChance = 0.5f;
-
-            // add rng ratings to pet device
-            // linear or biased?
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearDamage = GeneratePetDeviceRating(tier);
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearDamageResist = GeneratePetDeviceRating(tier);
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearCritDamage = GeneratePetDeviceRating(tier);
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearCritDamageResist = GeneratePetDeviceRating(tier);
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearCrit = GeneratePetDeviceRating(tier);
-            if (ratingChance > ThreadSafeRandom.Next(0.0f, 1.0f))
-                petDevice.GearCritResist = GeneratePetDeviceRating(tier);
-
-            petDevice.ItemWorkmanship = GetWorkmanship(tier);
-        }
-
-        public static int GeneratePetDeviceRating(int tier)
-        {
-            // thanks to morosity for this formula!
-            var baseRating = ThreadSafeRandom.Next(1, 10);
-
-            var chance = 0.4f + tier * 0.02f;
-            var rng = ThreadSafeRandom.Next(0.0f, 1.0f);
-            if (rng < chance)
-                baseRating += ThreadSafeRandom.Next(1, 10);
-
-            return baseRating;
-        }
-
-        private static WorldObject CreateRandomScroll(TreasureDeath profile, TreasureRoll roll = null)
-        {
-            // level 8 spell components shouldn't be in here,
-            // they should be associated with TreasureItemType.SpellComponent (peas)
-            if (roll == null && profile.Tier >= 7)
+            if (roll.IsClothing || roll.IsArmor || roll.IsWeapon || roll.IsJewelry || roll.IsDinnerware)
             {
-                // According to wiki, Tier 7 has a chance for level 8 spell components or level 7 spell scrolls (as does Tier 8)
-                // No indication of weighting in either direction, so assuming a 50/50 split
+                range.min = 0.9f;
+                range.max = 1.1f;
+            }
+            else if (!roll.IsGem)
+            {
+                log.Error($"RollSpellcraft({wo.Name}, {roll.ItemType}) - unknown item type");
+            }
+
+            var rng = ThreadSafeRandom.Next(range.min, range.max);
+
+            var spellcraft = (int)Math.Ceiling(maxSpellPower * rng);
+
+            // retail was capped at 370
+            spellcraft = Math.Min(spellcraft, 370);
+
+            return spellcraft;
+        }
+
+        /// <summary>
+        /// Returns the maximum power from the spells in item's spellbook
+        /// </summary>
+        public static int GetMaxSpellPower(WorldObject wo)
+        {
+            var maxSpellPower = 0;
+
+            if (wo.SpellDID != null)
+            {
+                var spell = new Server.Entity.Spell(wo.SpellDID.Value);
+
+                if (spell.Power > maxSpellPower)
+                    maxSpellPower = (int)spell.Power;
+            }
+
+            if (wo.Biota.PropertiesSpellBook != null)
+            {
+                foreach (var spellId in wo.Biota.PropertiesSpellBook.Keys)
+                {
+                    var spell = new Server.Entity.Spell(spellId);
+
+                    if (spell.Power > maxSpellPower)
+                        maxSpellPower = (int)spell.Power;
+                }
+            }
+            return maxSpellPower;
+        }
+
+        private static void AddActivationRequirements(WorldObject wo, TreasureRoll roll)
+        {
+            // ItemSkill/LevelLimit
+            TryMutate_ItemSkillLimit(wo, roll);
+
+            // Arcane Lore / ItemDifficulty
+            wo.ItemDifficulty = CalculateArcaneLore(wo, roll);
+        }
+
+        private static bool TryMutate_ItemSkillLimit(WorldObject wo, TreasureRoll roll)
+        {
+            if (!RollItemSkillLimit(roll))
+                return false;
+
+            wo.ItemSkillLevelLimit = wo.ItemSpellcraft + 20;
+
+            var skill = Skill.None;
+
+            if (roll.IsMeleeWeapon || roll.IsMissileWeapon)
+            {
+                skill = wo.WeaponSkill;
+            }
+            else if (roll.IsArmor)
+            {
                 var rng = ThreadSafeRandom.Next(0.0f, 1.0f);
 
                 if (rng < 0.5f)
                 {
-                    var wcid = RollLevel8SpellComp();
-
-                    return WorldObjectFactory.CreateNewWorldObject((uint)wcid);
+                    skill = Skill.MeleeDefense;
+                }
+                else
+                {
+                    skill = Skill.MissileDefense;
+                    wo.ItemSkillLevelLimit = (int)(wo.ItemSkillLevelLimit * 0.7f);
                 }
             }
-            var spellLevel = ScrollLevelChance.Roll(profile);
-
-            // todo: switch to SpellLevelProgression
-            var spellId = SpellId.Undef;
-            do
+            else
             {
-                var spellIdx = ThreadSafeRandom.Next(0, ScrollSpells.Table.Length - 1);
-
-                spellId = ScrollSpells.Table[spellIdx][spellLevel - 1];
-            }
-            while (spellId == SpellId.Undef);   // simple way of handling spells that start at level 3 (blasts, volleys)
-
-            var weenie = DatabaseManager.World.GetScrollWeenie((uint)spellId);
-
-            if (weenie == null)
-            {
-                log.DebugFormat("CreateRandomScroll for tier {0} and spellID of {1} returned null from the database.", profile.Tier, spellId);
-                return null;
+                log.Error($"RollItemSkillLimit({wo.Name}, {roll.ItemType}) - unknown item type");
+                return false;
             }
 
-            return WorldObjectFactory.CreateNewWorldObject(weenie.WeenieClassId);
+            wo.ItemSkillLimit = skill;
+            return true;
         }
 
-        private static int RollLevel8SpellComp()
+        private static bool RollItemSkillLimit(TreasureRoll roll)
         {
-            var rng = ThreadSafeRandom.Next(0, LootTables.Level8SpellComps.Length - 1);
+            if (roll.IsMeleeWeapon || roll.IsMissileWeapon)
+            {
+                return true;
+            }
+            else if (roll.IsArmor)
+            {
+                var rng = ThreadSafeRandom.Next(0.0f, 1.0f);
 
-            return LootTables.Level8SpellComps[rng];
+                return rng < 0.55f;
+            }
+            return false;
+        }
+
+        // previous method - replaces itemSkillLevelLimit w/ WieldDifficulty,
+        // and does not use treasure roll
+
+        private static int RollItemDifficulty(WorldObject wo)
+        {
+            // - # of spells on item
+            var num_spells = wo.Biota.PropertiesSpellBook.Count();
+
+            var spellAddonChance = num_spells * (20.0f / (num_spells + 2.0f));
+            var spellAddon = (float)ThreadSafeRandom.Next(1.0f, spellAddonChance) * num_spells;
+
+            // - # of epics / legendaries on item
+            var epicAddon = wo.EpicCantrips.Count > 0 ? ThreadSafeRandom.Next(1, 5) * wo.EpicCantrips.Count : 0;
+            var legAddon = wo.LegendaryCantrips.Count > 0 ? ThreadSafeRandom.Next(5, 10) * wo.LegendaryCantrips.Count : 0;
+
+            // wield difficulty - skill requirement
+            var wieldFactor = 0.0f;
+
+            if (wo.WieldDifficulty != null && wo.WieldRequirements == WieldRequirement.RawSkill)
+                wieldFactor = wo.WieldDifficulty.Value / 3.0f;
+
+            var itemDifficulty = wo.ItemSpellcraft.Value - wieldFactor;
+
+            if (itemDifficulty < 0)
+                itemDifficulty = 0;
+
+            return (int)Math.Floor(itemDifficulty + spellAddon + epicAddon + legAddon);
+        }
+
+        /// <summary>
+        /// Calculates the Arcane Lore requirement / ItemDifficulty
+        /// </summary>
+        private static int CalculateArcaneLore(WorldObject wo, TreasureRoll roll)
+        {
+            // spellcraft - (itemSkillLevelLimit / 2.0f) + creatureLifeEnchantments + cantrips
+
+            var spellcraft = wo.ItemSpellcraft.Value;
+
+            // - mutates 100% of the time for melee / missile weapons
+            // - mutates 55% of the time for armor
+            // - mutates 0% of the time for all other item types
+            var itemSkillLevelFactor = 0.0f;
+
+            if (wo.ItemSkillLevelLimit > 0)
+                itemSkillLevelFactor = wo.ItemSkillLevelLimit.Value / 2.0f;
+
+            var fArcane = spellcraft - itemSkillLevelFactor;
+
+            if (fArcane < 0)
+                fArcane = 0;
+
+            return (int)Math.Floor(fArcane + roll.ItemDifficulty);
         }
     }
 }
