+++ Source/ACE.Server/Factories/LootGenerationFactory_Test.cs
@@ -1,12 +1,9 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 
 using ACE.Database;
 using ACE.Database.Models.World;
-using ACE.Entity.Enum;
-using ACE.Server.Entity;
 using ACE.Server.WorldObjects;
 
 namespace ACE.Server.Factories
