+++ Source/ACE.Server/Factories/LootGenerationFactory_Weapon.cs
@@ -1,0 +1,44 @@
+using ACE.Common;
+using ACE.Database.Models.World;
+using ACE.Server.Factories.Tables;
+using ACE.Server.WorldObjects;
+
+namespace ACE.Server.Factories
+{
+    public static partial class LootGenerationFactory
+    {
+        private static WorldObject CreateWeapon(TreasureDeath profile, bool isMagical)
+        {
+            int chance = ThreadSafeRandom.Next(1, 100);
+
+            // Aligning drop ratio to better align with retail - HarliQ 11/11/19
+            // Melee - 42%
+            // Missile - 36%
+            // Casters - 22%
+
+            return chance switch
+            {
+                var rate when (rate < 43) => CreateMeleeWeapon(profile, isMagical),
+                var rate when (rate > 42 && rate < 79) => CreateMissileWeapon(profile, isMagical),
+                _ => CreateCaster(profile, isMagical),
+            };
+        }
+
+        private static float RollWeaponSpeedMod(TreasureDeath treasureDeath)
+        {
+            var qualityLevel = QualityChance.Roll(treasureDeath);
+
+            if (qualityLevel == 0)
+                return 1.0f;    // no bonus
+
+            var rng = (float)ThreadSafeRandom.Next(-0.025f, 0.025f);
+
+            // min/max range: 67.5% - 100%
+            var weaponSpeedMod = 1.0f - (qualityLevel * 0.025f + rng);
+
+            //Console.WriteLine($"WeaponSpeedMod: {weaponSpeedMod}");
+
+            return weaponSpeedMod;
+        }
+    }
+}
